+++ index.js
@@ -48,19 +48,29 @@ class DabKnex extends Dab {
     if (this._.isPlainObject(coll)) {
       coll.srcAttribName = 'table'
       coll.srcAttribId = 'id'
-      coll.srcAttribIdType = 'string'
+      coll.srcAttribIdType = coll.srcAttribIdType || 'integer'
     }
     return new Promise((resolve, reject) => {
       super.createCollection(coll)
         .then(result => {
           this.setClient()
-          let rebuild = params.withSchema && !this._.isEmpty(this.collection[coll.name].attributes)
+          let rebuild = params.rebuild && !this._.isEmpty(this.collection[coll.name].attributes)
           if (!rebuild) return resolve(result)
           return this.client.schema.dropTableIfExists(coll.name)
         })
         .then(result => {
+          const cl = this.collection[coll.name]
           return this.client.schema.createTable(coll.name, table => {
-            this._.forOwn(this.collection[coll.name].attributes, (f, id) => {
+            // pk
+            const pk = this._.find(cl.attributes, { primaryKey: true })
+            if (cl.srcAttribIdType === 'string') {
+              table.string(coll.srcAttribId, pk.length || 255).notNullable().primary()
+            } else {
+              table.increments(coll.srcAttribId).notNullable().primary()
+            }
+            // columns
+            this._.forOwn(cl.attributes, (f, id) => {
+              if (id === cl.srcAttribId) return
               let column
               switch (f.type) {
                 case 'string':
@@ -69,18 +79,35 @@ class DabKnex extends Dab {
                 case 'datetime':
                   column = table.dateTime(id)
                   break
+                case 'object':
+                  column = table.specificType(id, 'json[]')
+                  break
+                case 'array':
+                  const subtype = f.subTypeOf || 'json'
+                  column = table.specificType(id, subtype + '[]')
+                  break
                 default:
                   column = table[f.type](id)
               }
               if (f.validator.required) column.notNullable()
               if (f.default !== undefined) column.defaultTo(f.default)
             })
+            // indexes
+            this._.forOwn(this.collection[coll.name].indexes, (v, k) => {
+              if (v.unique) {
+                table.unique(v.column, coll.name + '_' + k)
+              } else {
+                table.index(v.column, coll.name + '_' + k)
+              }
+            })
           })
         })
         .then(result => {
           resolve(true)
         })
-        .catch(err => reject(this._error(err)))
+        .catch(err => {
+          reject(this._error(err))
+        })
     })
   }
 
@@ -90,7 +117,7 @@ class DabKnex extends Dab {
       super.renameCollection(oldName, newName)
         .then(result => {
           this.setClient()
-          let rebuild = params.withSchema && !this._.isEmpty(this.collection[newName].attributes)
+          let rebuild = params.rebuild && !this._.isEmpty(this.collection[newName].attributes)
           if (!rebuild) return resolve(result)
           return this.client.schema.renameTable(oldName, newName)
         })
@@ -103,7 +130,7 @@ class DabKnex extends Dab {
 
   removeCollection (name, params) {
     params = params || {}
-    let rebuild = params.withSchema && this.collection[name] && !this._.isEmpty(this.collection[name].attributes)
+    let rebuild = params.rebuild && this.collection[name] && !this._.isEmpty(this.collection[name].attributes)
     return new Promise((resolve, reject) => {
       super.removeCollection(name)
         .then(result => {
@@ -141,6 +168,7 @@ class DabKnex extends Dab {
             data: []
           }
           result.forEach((d, i) => {
+            d = this._transformTo(d, params)
             data.data.push(this.convert(d, { collection: params.collection }))
           })
           resolve(data)
@@ -150,8 +178,10 @@ class DabKnex extends Dab {
   }
 
   _findOne (id, params, callback) {
+    const coll = this.collection[params.collection]
+    if (!coll) throw new Error('Collection not found')
     this.client(params.collection)
-      .where('id', id)
+      .where(coll.srcAttribId, id)
       .then(data => {
         if (data.length === 0) {
           return callback(null, {
@@ -161,7 +191,7 @@ class DabKnex extends Dab {
         }
         data = {
           success: true,
-          data: data[0]
+          data: this._transformTo(data[0], params)
         }
         callback(null, data)
       })
@@ -188,10 +218,38 @@ class DabKnex extends Dab {
     })
   }
 
+  _transformFrom (body, params) {
+    if (this.options.client !== 'sqlite3') return body
+    const newBody = this._.cloneDeep(body)
+    const coll = this.collection[params.collection].attributes
+    if (!coll) throw new Error('Collection not found')
+    this._.forOwn(newBody, (v, k) => {
+      if (coll[k] && ['array', 'object'].indexOf(coll[k].type) > -1) newBody[k] = JSON.stringify(v)
+    })
+    return newBody
+  }
+
+  _transformTo (body, params) {
+    if (this.options.client !== 'sqlite3') return body
+    const newBody = this._.cloneDeep(body)
+    const coll = this.collection[params.collection].attributes
+    if (!coll) throw new Error('Collection not found')
+    this._.forOwn(newBody, (v, k) => {
+      if (!this._.isEmpty(v) && coll[k] && ['array', 'object'].indexOf(coll[k].type) > -1) newBody[k] = JSON.parse(v)
+    })
+    return newBody
+  }
+
   _create (body, params, callback) {
-    this.client(params.collection).insert(body, 'id')
+    const coll = this.collection[params.collection]
+    if (!coll) throw new Error('Collection not found')
+    if (!this._.has(body, coll.srcAttribId) && coll.srcAttribIdType === 'string') {
+      body[coll.srcAttribId] = this.nanoid(13)
+    }
+    body = this._transformFrom(body, params)
+    this.client(params.collection).insert(body, coll.srcAttribId)
       .then(result => {
-        let id = this._.isString(body.id) ? body.id : result[0]
+        let id = this._.isString(body[coll.srcAttribId]) ? body[coll.srcAttribId] : result[0]
         this._findOne(id, params, callback)
       })
       .catch(err => {
@@ -204,10 +262,12 @@ class DabKnex extends Dab {
 
   create (body, params) {
     [params, body] = this.sanitize(params, body)
+    const coll = this.collection[params.collection]
+    if (!coll) throw new Error('Collection not found')
     this.setClient(params)
     return new Promise((resolve, reject) => {
-      if (body.id) {
-        this._findOne(body.id, params, (e, result) => {
+      if (body[coll.srcAttribId]) {
+        this._findOne(body[coll.srcAttribId], params, (e, result) => {
           if (result.success) return reject(new Error('Document already exists'))
           this._create(body, params, (e, result) => {
             if (!result.success) return reject(result.err)
@@ -227,14 +287,17 @@ class DabKnex extends Dab {
 
   update (id, body, params) {
     [params, body] = this.sanitize(params, body)
+    const coll = this.collection[params.collection]
+    if (!coll) throw new Error('Collection not found')
     this.setClient(params)
-    body = this._.omit(body, ['id'])
+    body = this._.omit(body, [coll.srcAttribId])
+    body = this._transformFrom(body, params)
     return new Promise((resolve, reject) => {
       this._findOne(id, params, (e, result) => {
         if (!result.success) return reject(result.err)
         let source = result.data
         this.client(params.collection)
-          .where('id', id)
+          .where(coll.srcAttribId, id)
           .update(body)
           .then(result => {
             this._findOne(id, params, (e, result) => {
@@ -254,12 +317,14 @@ class DabKnex extends Dab {
 
   remove (id, params) {
     [params] = this.sanitize(params)
+    const coll = this.collection[params.collection]
+    if (!coll) throw new Error('Collection not found')
     this.setClient(params)
     return new Promise((resolve, reject) => {
       this._findOne(id, params, (e, result) => {
         if (!result.success) return reject(result.err)
         let source = result.data
-        this.client(params.collection).where('id', id).del()
+        this.client(params.collection).where(coll.srcAttribId, id).del()
           .then(result => {
             let data = {
               success: true
@@ -274,35 +339,37 @@ class DabKnex extends Dab {
 
   bulkCreate (body, params) {
     [params, body] = this.sanitize(params, body)
+    const coll = this.collection[params.collection]
+    if (!coll) throw new Error('Collection not found')
     this.setClient(params)
     return new Promise((resolve, reject) => {
       if (!this._.isArray(body)) return reject(new Error('Requires an array'))
 
       this._.each(body, (b, i) => {
-        if (!b.id) b.id = this.uuid()
-        body[i] = b
+        if (!b[coll.srcAttribId]) b[coll.srcAttribId] = this.nanoid(13)
+        body[i] = this._transformFrom(b, params)
       })
-      const keys = this._(body).map('id').value()
+      const keys = this._(body).map(coll.srcAttribId).value()
 
-      this.client.select('id').from(params.collection)
-        .whereIn('id', keys).asCallback((err, docs) => {
+      this.client.select(coll.srcAttribId).from(params.collection)
+        .whereIn(coll.srcAttribId, keys).asCallback((err, docs) => {
           if (err) return reject(this._error(err))
-          let info = this._.map(docs, 'id')
+          let info = this._.map(docs, coll.srcAttribId)
           let newBody = this._.clone(body)
           this._.pullAllWith(newBody, info, (i, x) => {
-            return i.id === x
+            return i[coll.srcAttribId] === x
           })
           async.mapSeries(newBody, (b, cb) => {
-            this.client(params.collection).insert(b, 'id').asCallback((err, result) => {
-              cb(null, err ? { id: b.id, message: err.message } : null)
+            this.client(params.collection).insert(b, coll.srcAttribId).asCallback((err, result) => {
+              cb(null, err ? { id: b[coll.srcAttribId], message: err.message } : null)
             })
           }, (err, result) => {
             if (err) return reject(err)
             let ok = 0
             let status = []
             this._.each(body, (r, i) => {
-              let stat = { success: info.indexOf(r.id) === -1 }
-              stat.id = r.id
+              let stat = { success: info.indexOf(r[coll.srcAttribId]) === -1 }
+              stat[coll.srcAttribId] = r[coll.srcAttribId]
               if (!stat.success) stat.message = 'Document already exists'
               else ok++
               status.push(stat)
@@ -324,42 +391,44 @@ class DabKnex extends Dab {
 
   bulkUpdate (body, params) {
     [params, body] = this.sanitize(params, body)
+    const coll = this.collection[params.collection]
+    if (!coll) throw new Error('Collection not found')
     this.setClient(params)
     return new Promise((resolve, reject) => {
       if (!this._.isArray(body)) return reject(new Error('Requires an array'))
       this._.each(body, (b, i) => {
-        if (!b.id) b.id = this.uuid()
-        body[i] = b
+        if (!b[coll.srcAttribId]) b[coll.srcAttribId] = this.nanoid(13)
+        body[i] = this._transformFrom(b, params)
       })
-      const keys = this._(body).map('id').value()
+      const keys = this._(body).map(coll.srcAttribId).value()
 
-      this.client.select('id').from(params.collection)
-        .whereIn('id', keys).asCallback((err, docs) => {
+      this.client.select(coll.srcAttribId).from(params.collection)
+        .whereIn(coll.srcAttribId, keys).asCallback((err, docs) => {
           if (err) return reject(this._error(err))
-          let info = this._.map(docs, 'id')
+          let info = this._.map(docs, coll.srcAttribId)
           let newBody = []
 
           this._.each(body, b => {
-            if (info.indexOf(b.id) > -1) newBody.push(b)
+            if (info.indexOf(b[coll.srcAttribId]) > -1) newBody.push(b)
           })
           async.mapSeries(newBody, (b, cb) => {
-            this.client(params.collection).where('id', b.id).update(this._.omit(b, 'id')).asCallback((err, result) => {
-              cb(null, err ? { id: b.id, message: err.message } : null)
+            this.client(params.collection).where(coll.srcAttribId, b[coll.srcAttribId]).update(this._.omit(b, coll.srcAttribId)).asCallback((err, result) => {
+              cb(null, err ? { id: b[coll.srcAttribId], message: err.message } : null)
             })
           }, (err, result) => {
             if (err) return reject(err)
             let ok = 0
             let status = []
             this._.each(body, (r, i) => {
-              let stat = { success: info.indexOf(r.id) > -1 }
+              let stat = { success: info.indexOf(r[coll.srcAttribId]) > -1 }
               if (stat.success) {
-                let rec = this._.find(result, { id: r.id })
+                let rec = this._.find(result, { id: r[coll.srcAttribId] })
                 if (rec) {
                   stat.success = false
                   stat.message = rec.message
                 }
               }
-              stat.id = r.id
+              stat[coll.srcAttribId] = r[coll.srcAttribId]
               if (!stat.success && !stat.message) stat.message = 'Document not found'
               else ok++
               status.push(stat)
@@ -381,23 +450,25 @@ class DabKnex extends Dab {
 
   bulkRemove (body, params) {
     [params, body] = this.sanitize(params, body)
+    const coll = this.collection[params.collection]
+    if (!coll) throw new Error('Collection not found')
     this.setClient(params)
     return new Promise((resolve, reject) => {
       if (!this._.isArray(body)) return reject(new Error('Requires an array'))
       this._.each(body, (b, i) => {
-        body[i] = b || this.uuid()
+        body[i] = b || this.nanoid(13)
       })
 
-      this.client.select('id').from(params.collection)
-        .whereIn('id', body).asCallback((err, docs) => {
+      this.client.select(coll.srcAttribId).from(params.collection)
+        .whereIn(coll.srcAttribId, body).asCallback((err, docs) => {
           if (err) return reject(this._error(err))
-          let info = this._.map(docs, 'id')
+          let info = this._.map(docs, coll.srcAttribId)
           let newBody = []
           this._.each(body, b => {
             if (info.indexOf(b) > -1) newBody.push(b)
           })
           async.mapSeries(newBody, (b, cb) => {
-            this.client(params.collection).where('id', b).del().asCallback((err, result) => {
+            this.client(params.collection).where(coll.srcAttribId, b).del().asCallback((err, result) => {
               cb(null, err ? { id: b, message: err.message } : null)
             })
           }, (err, result) => {
@@ -406,7 +477,7 @@ class DabKnex extends Dab {
             let status = []
             this._.each(body, (r, i) => {
               let stat = { success: info.indexOf(r) > -1 }
-              stat.id = r
+              stat[coll.srcAttribId] = r
               if (!stat.success) stat.message = 'Document not found'
               else ok++
               status.push(stat)